requirements.txt
# Clean dependencies for ForeTrip Weather API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
aiohttp==3.9.1